feat structure: moved Metric class to metric_calculator.py
