--- src/cpp_stats/metrics/metric.py
@@ -1,14 +1,0 @@
-class Metric:
-    '''
-    Class that stores metric name and value.
-    '''
-
-    def __add__(self, other):
-        pass
-
-    def get(self):
-        '''
-        Returns metric value. Should be called since metric can be summarization of many files,
-        namespaces, classes and etc.
-        '''
-        pass
