+++ src/cpp_stats/metrics/metric_calculator.py
@@ -2,7 +2,20 @@ from pathlib import Path
 
 import clang.cindex
 
-from cpp_stats.metrics.metric import Metric
+class Metric:
+    '''
+    Class that stores metric name and value.
+    '''
+
+    def __add__(self, other):
+        pass
+
+    def get(self):
+        '''
+        Returns metric value. Should be called since metric can be summarization of many files,
+        namespaces, classes and etc.
+        '''
+        pass
 
 class BasicMetricCalculator:
     '''
